Correcao de TIpos de Dados

+++ src/main/java/application/model/v1/Cartao.java
@@ -11,7 +11,7 @@ public class Cartao {
 	public String nomeTitular;
 	public String cpfTitular;
 	public Boolean bloqueado;
-	public BigDecimal saldo;
+	public Double saldo;
 	public String numero;
